RoleSchool adapter modified to include RS-id
- To extend our web service with the course handling, it should be
mandatory to include the id of a role school when saving it in a json
object. Before there was no id connected to our role school json
objects, but now it is modified in the adapter to include it, so we can
assign and remove specific roleschool-objects to specific courses,
thanks to their distinct ids.

Commit: Christopher